Move derive_has_fields
