+++ crates/macros/cgp-macro-core/src/types/cgp_data/derive_has_fields/derive_struct.rs
@@ -2,10 +2,10 @@ use proc_macro2::TokenStream;
 use quote::quote;
 use syn::{ItemImpl, ItemStruct, parse2};
 
-use crate::derive_has_fields::from_fields_struct::derive_from_fields_for_struct;
-use crate::derive_has_fields::product::item_fields_to_product_type;
-use crate::derive_has_fields::to_fields_ref_struct::derive_to_fields_ref_for_struct;
-use crate::derive_has_fields::to_fields_struct::derive_to_fields_for_struct;
+use crate::types::cgp_data::{
+    derive_from_fields_for_struct, derive_to_fields_for_struct, derive_to_fields_ref_for_struct,
+    item_fields_to_product_type,
+};
 
 pub fn derive_has_fields_impls_from_struct(item_struct: &ItemStruct) -> syn::Result<Vec<ItemImpl>> {
     let struct_name = &item_struct.ident;
