+++ crates/macros/cgp-macro-core/src/types/cgp_data/derive_has_fields/from_fields_struct.rs
