+++ crates/macros/cgp-macro-core/src/types/cgp_data/derive_has_fields/sum.rs
@@ -1,11 +1,11 @@
-use cgp_macro_core::types::field::Symbol;
 use proc_macro2::TokenStream;
 use quote::quote;
 use syn::punctuated::Punctuated;
 use syn::token::Comma;
 use syn::{Type, Variant, parse2};
 
-use crate::derive_has_fields::product::item_fields_to_product_type;
+use crate::types::cgp_data::item_fields_to_product_type;
+use crate::types::field::Symbol;
 
 pub fn variants_to_sum_type(
     variants: &Punctuated<Variant, Comma>,
