+++ crates/macros/cgp-macro-core/src/types/cgp_data/derive_has_fields/to_fields_struct.rs
