+++ crates/macros/cgp-macro-core/src/types/cgp_data/mod.rs
@@ -1,3 +1,5 @@
+mod derive_has_fields;
 mod record;
 
+pub use derive_has_fields::*;
 pub use record::*;
